(many) Implement bigint type (#238)
A large part of this was already in place:

- The result of the expontential operator with int operands produced a
  `BigInteger` value.
- The `Scanner` would create a `BigInteger` if a constant expression was
  encountered where the value would not fit into `int` or `long`
- Implementation of arithmetic operators (-, + etc) was mostly in place
  already.

What this PR adds:

- Support for explicitly declaring variables of `bigint` type.
- Integration tests to ensure the new type works in various scenarios

+++ src/Perlang.Common/Extensions/TypeExtensions.cs
@@ -15,7 +15,7 @@ namespace Perlang.Extensions
                 { } when type == typeof(Double) => "double",
                 { } when type == typeof(Int32) => "int",
                 { } when type == typeof(Int64) => "long",
-                { } when type == typeof(BigInteger) => "BigInteger", // TODO: Should we make this bigint and support it as a special type as the others?
+                { } when type == typeof(BigInteger) => "bigint",
                 { } when type == typeof(NullObject) => "null",
                 { } when type == typeof(String) => "string",
                 { } when type.IsAssignableTo(typeof(IPerlangFunction)) => "function",
